+++ tests/ci/docker_images_check.py
@@ -10,6 +10,7 @@ from s3_helper import S3Helper
 from pr_info import PRInfo
 from get_robot_token import get_best_robot_token, get_parameter_from_ssm
 from upload_result_helper import upload_results
+from commit_status_helper import get_commit
 
 NAME = "Push to Dockerhub (actions)"
 
@@ -149,11 +150,6 @@ def process_test_results(s3_client, test_results, s3_path_prefix):
         processed_test_results.append((test_name, status))
     return overall_status, processed_test_results
 
-def get_commit(gh, commit_sha):
-    repo = gh.get_repo(os.getenv("GITHUB_REPOSITORY", "ClickHouse/ClickHouse"))
-    commit = repo.get_commit(commit_sha)
-    return commit
-
 if __name__ == "__main__":
     logging.basicConfig(level=logging.INFO)
     repo_path = os.getenv("GITHUB_WORKSPACE", os.path.abspath("../../"))
@@ -201,12 +197,11 @@ if __name__ == "__main__":
 
     url = upload_results(s3_helper, pr_info.number, pr_info.sha, test_results, [], NAME)
 
-    gh = Github(get_best_robot_token())
-    commit = get_commit(gh, pr_info.sha)
-    commit.create_status(context=NAME, description=description, state=status, target_url=url)
-
     with open(os.path.join(temp_path, 'changed_images.json'), 'w') as images_file:
         json.dump(result_images, images_file)
 
     print("::notice ::Report url: {}".format(url))
     print("::set-output name=url_output::\"{}\"".format(url))
+    gh = Github(get_best_robot_token())
+    commit = get_commit(gh, pr_info.sha)
+    commit.create_status(context=NAME, description=description, state=status, target_url=url)
