+++ tests/ci/functional_test_check.py
@@ -5,7 +5,6 @@ import logging
 import subprocess
 import os
 import json
-import time
 import sys
 
 from github import Github
@@ -16,12 +15,8 @@ from pr_info import PRInfo
 from build_download_helper import download_all_deb_packages
 from upload_result_helper import upload_results
 from docker_pull_helper import get_image_with_version
+from commit_status_helper import post_commit_status, get_commit
 
-
-def get_commit(gh, commit_sha):
-    repo = gh.get_repo(os.getenv("GITHUB_REPOSITORY", "ClickHouse/ClickHouse"))
-    commit = repo.get_commit(commit_sha)
-    return commit
 
 def get_image_name(check_name):
     if 'stateless' in check_name.lower():
@@ -161,6 +156,6 @@ if __name__ == "__main__":
     s3_helper = S3Helper('https://s3.amazonaws.com')
     state, description, test_results, additional_logs = process_results(result_path, server_log_path)
     report_url = upload_results(s3_helper, pr_info.number, pr_info.sha, test_results, [run_log_path] + additional_logs, check_name)
+
     print(f"::notice ::Report url: {report_url}")
-    commit = get_commit(gh, pr_info.sha)
-    commit.create_status(context=check_name, description=description, state=state, target_url=report_url)
+    post_commit_status(gh, pr_info.sha, check_name, description, state, report_url)
