+++ tests/ci/integration_test_check.py
@@ -4,7 +4,6 @@ import os
 import logging
 import sys
 import json
-import time
 import subprocess
 import csv
 
@@ -16,6 +15,7 @@ from pr_info import PRInfo
 from build_download_helper import download_all_deb_packages
 from upload_result_helper import upload_results
 from docker_pull_helper import get_images_with_versions
+from commit_status_helper import post_commit_status
 
 DOWNLOAD_RETRIES_COUNT = 5
 
@@ -75,11 +75,6 @@ def get_images_with_versions(images_path):
         return result
     else:
         return {image: 'latest' for image in IMAGES}
-
-def get_commit(gh, commit_sha):
-    repo = gh.get_repo(os.getenv("GITHUB_REPOSITORY", "ClickHouse/ClickHouse"))
-    commit = repo.get_commit(commit_sha)
-    return commit
 
 def process_results(result_folder):
     test_results = []
@@ -171,5 +166,4 @@ if __name__ == "__main__":
     s3_helper = S3Helper('https://s3.amazonaws.com')
     report_url = upload_results(s3_helper, pr_info.number, pr_info.sha, test_results, [output_path_log] + additional_logs, check_name, False)
     print(f"::notice ::Report url: {report_url}")
-    commit = get_commit(gh, pr_info.sha)
-    commit.create_status(context=check_name, description=description, state=state, target_url=report_url)
+    post_commit_status(gh, pr_info.sha, check_name, description, state, report_url)
