+++ tests/ci/unit_tests_check.py
@@ -14,13 +14,9 @@ from pr_info import PRInfo
 from build_download_helper import download_unit_tests
 from upload_result_helper import upload_results
 from docker_pull_helper import get_image_with_version
+from commit_status_helper import post_commit_status
 
 IMAGE_NAME = 'clickhouse/unit-test'
-
-def get_commit(gh, commit_sha):
-    repo = gh.get_repo(os.getenv("GITHUB_REPOSITORY", "ClickHouse/ClickHouse"))
-    commit = repo.get_commit(commit_sha)
-    return commit
 
 def get_test_name(line):
     elements = reversed(line.split(' '))
@@ -137,5 +133,4 @@ if __name__ == "__main__":
     state, description, test_results, additional_logs = process_result(test_output)
     report_url = upload_results(s3_helper, pr_info.number, pr_info.sha, test_results, [run_log_path] + additional_logs, check_name)
     print(f"::notice ::Report url: {report_url}")
-    commit = get_commit(gh, pr_info.sha)
-    commit.create_status(context=check_name, description=description, state=state, target_url=report_url)
+    post_commit_status(gh, pr_info.sha, check_name, description, state, report_url)
